Move mcp specifics into BaseAgent
